// src/api/generated/models/ticketVisibility.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Controls whether a ticket is visible in the public community list.
 */
export type TicketVisibility =
  (typeof TicketVisibility)[keyof typeof TicketVisibility]

export const TicketVisibility = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE',
} as const